pyproject.toml
[tool.poetry]
name = "photocurate"
version = "0.1.0"
description = "Multi-tenant SaaS platform for photographers — AI-powered photo scoring, client galleries, and cloud delivery."
authors = ["PhotoCurate Team"]
readme = "README.md"
packages = [{include = "photocurate", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.0"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
pyvips = "^2.2.0"
Pillow = "^10.3.0"
opencv-python-headless = "^4.9.0"
imagehash = "^4.3.0"
numpy = "^1.26.0"
azure-storage-blob = {version = "^12.20.0", optional = true}
azure-servicebus = {version = "^7.12.0", optional = true}
boto3 = {version = "^1.34.0", optional = true}
nats-py = {version = "^2.7.0", optional = true}

[tool.poetry.extras]
azure = ["azure-storage-blob", "azure-servicebus"]
selfhost = ["boto3", "nats-py"]
all = ["azure-storage-blob", "azure-servicebus", "boto3", "nats-py"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
